refactor(app): reuse shared logging target
